seed/requirements.txt
psycopg2==2.7.7
python-dotenv==0.18.0
simfin==0.8.1
SQLAlchemy==1.4.2